move diff utils
